Split test-utils.test.ts: extract test-compat describe block
test-utils.test.ts (1380 lines) exceeded the 1200-line limit.

- test-utils.test.ts (986 lines): tests for all #test-utils helper functions
  (createTestDb, createTestEvent, loginAsAdmin, withSetting, etc.)
- test-compat.test.ts (396 lines): Deno test framework compatibility tests
  (expect matchers, FakeTime, spy/stub, fn(), group helpers)
